+++ lang/en/tool_vault.php
@@ -124,6 +124,7 @@ $string['error_restoreprecheckfailed'] = 'Error occurred while executing restore
 $string['error_restoresnotallowed'] = 'Restores are not allowed on this site';
 $string['error_serverreturnednodata'] = 'Server returned no data';
 $string['error_unabletorunprecheck'] = 'Unable to run pre-check {$a}';
+$string['error_usecli'] = 'You can not perform site backup and restore from this page, please use command-line interface.';
 $string['excludetablefrombackup'] = 'Add "{$a}" to excluded tables setting.';
 $string['expandlogs'] = 'Expand logs';
 $string['forgetapikey'] = 'Forget API key';
@@ -207,6 +208,8 @@ For most plugins, you can include them in the backup and uninstall them after re
 Note that this will only exclude data in the plugin's own tables, settings, associated files, scheduled tasks and other known common types of plugin-related data. It may not be accurate for complicated plugins or plugins with dependencies.";
 $string['settings_backupexcludetables'] = 'Exclude tables';
 $string['settings_backupexcludetables_desc'] = "Vault will back up all tables that start with the prefix '{\$a->prefix}' even if they are not listed in xmldb schemas of the core or installed plugins.<br>Here you can list extra tables that should not be backed up. Use an asterisk ('*') to exclude all extra tables.<br>You must include the table prefix when excluding tables.";
+$string['settings_clionly'] = 'Allow to perform site backup and restore from CLI only';
+$string['settings_clionly_desc'] = 'Checking this will hide the "Vault - Site migration" from the site administration and prevent access to backup and restore. ';
 $string['settings_desc'] = "Vault plugin settings where you can configure what to exclude during backup or preserve during restore.";
 $string['settings_header'] = 'Vault - Site migration settings';
 $string['settings_headerbackup'] = 'Backup settings';