+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to the tool_vault plugin will be documented in this file.
 ## Unreleased
 
 ### Added
-- Improvements to the CLI backup/restore
+- Improvements to the CLI backup/restore, allow to run backup pre-check
+- Setting to disable web access to tool_vault and execute from CLI only
 
 ## [1.8] - 2024-04-30